Revert "fairer benchmark with -O2 for everyone"
This reverts commit 904fa69df4b296cdb86e34407d7da48b42167ac0.

Instead of making it a fair benchmark by using -O2 for everyone, I
accidentally made it unfair. It did cause all the countDown
implementations to be compiled with -O2, but it also caused the
too-fast-too-free library itself to be compiled with -O2, which isn't
how freer-simple and mtl are compiled.