remove type application in example
I think the thought here is to assume the plugin is used. If that's the
case, at least on ghc 8.8.3, the type application isn't needed on catch
with polysemy-plugin enabled.

Added comment about needing type application without the plugin in line
with the reason for this PR for users who might just copy/paste to try polysemy.